requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.3
django-filter==2.1.0
djangorestframework==3.9.0
factory-boy==2.11.1
Faker==1.0.4
idna==2.8
Pillow==5.3.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyJWT==1.6.4
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.7
redis==3.0.1
requests==2.21.0
# rollbar==0.14.7
six==1.12.0
text-unidecode==1.2
urllib3==1.24.1
